More Details
Added page for all employee attributes
